+++ CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Added
+### Changed
+* Loosen requirement for requests to 2.0 or higher. 
+### Deprecated
+### Fixed
+### Removed
+### Security
+
+
 ## [1.1.2] - 2021-09-21
 ### Changed
 * Improve request exception messages and include exception stacktraces.
@@ -66,6 +76,7 @@ Version increased to avoid conflicts with old packages on PyPI.
 Initial version.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-python/compare/v1.1.2...HEAD
 [1.1.2]: https://github.com/DeepLcom/deepl-python/compare/v1.1.1...v1.1.2
 [1.1.1]: https://github.com/DeepLcom/deepl-python/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/DeepLcom/deepl-python/compare/v1.0.1...v1.1.0