+++ CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+* Fix bug that caused User-Agent header to be omitted from HTTP requests. 
+### Security
+
+
 ## [1.2.0] - 2021-10-07
 ### Added
 * Add `Translator.get_glossary_languages()` to query language pairs supported for glossaries.
@@ -85,7 +95,8 @@ Version increased to avoid conflicts with old packages on PyPI.
 Initial version.
 
 
-[1.2.0]: https://github.com/DeepLcom/deepl-python/compare/v1.1.3...v1.2.0 
+[Unreleased]: https://github.com/DeepLcom/deepl-python/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/DeepLcom/deepl-python/compare/v1.1.3...v1.2.0
 [1.1.3]: https://github.com/DeepLcom/deepl-python/compare/v1.1.2...v1.1.3
 [1.1.2]: https://github.com/DeepLcom/deepl-python/compare/v1.1.1...v1.1.2
 [1.1.1]: https://github.com/DeepLcom/deepl-python/compare/v1.1.0...v1.1.1